Correctly set default zone for connection
When setting the zone for a connection to the default using `nmcli`, the
value for the connection.zone should be '' (empty string) (eg `nmcli
connection modify Starbucks connection.zone ''`). However, the DBus
GetSettings and Update methods use JSON to represent the settings. Using
'' for connection.zone when calling the DBus Update method sets the zone
for the connection to '' instead of the default zone. Instead,
connection.zone should simply be excluded from the JSON.