Give up on GNOME Shell Extension
If we wanted to make this a GNOME Shell Extension, the only thing the
extension itself would do would be to launch the GTK app. While EGO is
great for software distribution, I've [read][1] "Extensions are really
meant as a way to modify the compositor, not as an alternative app
distribution system". So it seems that we should abandon any effort to
distribute Zone Defense via EGO.

[1]: https://discourse.gnome.org/t/set-application-display-name-and-icon-without-an-installed-desktop-file/6095/14