Use consistent name for network connection ID
This will make it easier to track what's going on. We'll generally use
'connection ID' (eg `connectionId` in JavaScript) to refer to the
network connection ID. This is the name given by NetworkManager in the
D-Bus `GetSettings` and `Update` methods as well as `nmcli`.

However, when we present the connection ID to the user, we will use the
term 'network'. This will hopefully be more understandable for users.