Move directory creation into `init` method
Moving this code earlier allows us to fail fast. If we can't create the
directory, we won't be able to save, and we should just quit. We still
attempt to create the directory again just prior to use just in case the
user has done something to it since the program launched.

Also handle the case where you can't write to connection-ids-seen.json.