Fix race condition in startup
By calling `this.#connectionIdsSeen.init()` from the `main.js`
`constructor`, we were effectively still waiting for `ConnectionIdsSeen`
initialization to finish while resuming the rest of the application. In
actual use, this didn't seem to be a problem, but running the test
exposed it since it would attempt to use `ConnectionIdsSeen` immediatly
after instantiation.