Support eslint
I had some hacked something together earlier, but this is at least a
little better. Based on the [GJS Style Guide][1].

`networkState.js` looks wildly different, but we just changed the order
of the classes to keep the linter happy.

We'll include a minimal package.json, and will ignore package-lock.json.
This is at least partially based on some projects by [Andy Holmes][2].

[1]: https://gjs.guide/guides/gjs/style-guide.html
[2]: https://github.com/andyholmes