Include an action for handling errors
We'll include an action in main.js that other classes can activate in
the event of an error. This will help in cases where we want to notify
the user of a problem.

We may do something different in the future, but for now the action
always generates a notification.